Ifdeffed out multilingual support for the time being since none of our users use it and the voice prompts have fallen behind a long way. This saves us some memory and resolves the nasty side effect of fm calibration failing on boot due to memory usage. This removes the language menu from the main menu. To re-enable,
The calibration issue was resolvable by calling trxUpdateAt1846Calibration() e.g. on long hold sk2, not acceptable but proved that the updating of calibration failed on initial channel load. It could always be resolve by arrowing away and back to the channel or frequency, or by activating monitor and then releasing.
The issue only happens on FM tx and you'd notice it because if you listened to your tx, it would be heavily distorted or not work at all.